chore: refactor Optional utils to common place
